+++ CHANGELOG.md
@@ -10,6 +10,10 @@ Minor releases include updated stdlib stubs from typeshed.
 Pyxel doesn't use SemVer anymore, since most minor releases have at least minor backward incompatible changes.
 
 ## UNRELEASED
+This release brings a number of bugfixes, documentation and improvements.
+
+The following new models were added:
+* `load_star_map`.
 
 ### Core
 
@@ -22,8 +26,9 @@ Pyxel doesn't use SemVer anymore, since most minor releases have at least minor 
   (See [!728](https://gitlab.com/esa/pyxel/-/merge_requests/728)).
 
 ### Models
-* Add first model for group 'Scene Generation'.
-  (See [!715](https://gitlab.com/esa/pyxel/-/merge_requests/715)).
+* Add first model for group `load_star_map` in `Scene Generation`.
+  (See [!715](https://gitlab.com/esa/pyxel/-/merge_requests/715)
+  and [!716](https://gitlab.com/esa/pyxel/-/merge_requests/716)).
 
 ### Others